% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bias.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Calculate bias after simulations}
\usage{
calculate_bias(simdf)
}
\arguments{
\item{simdf}{Data frame with simulation results, as produced by \code{\link{run_sims}}.}
}
\value{
A tidy data frame with simulation details and resulting bias for each method tested.
}
\description{
Calculate bias after simulations
}
