% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_bayesmix}
\alias{fit_bayesmix}
\title{Fit Bayesian mixed model (brms)}
\usage{
fit_bayesmix(simdata, delta = 0.95)
}
\arguments{
\item{simdata}{List generated by \code{\link{simul_data}}.}

\item{delta}{adapt_delta parameter for brm}
}
\value{
A data frame with 4 columns (intercept estimate and standard error, and slope estimate and standard error), and as many rows as taxa.
}
\description{
Fit Bayesian mixed model (brms)
}
