% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_lump}
\alias{fit_lump}
\title{Fit lump model}
\usage{
fit_lump(simdata)
}
\arguments{
\item{simdata}{List generated by \code{\link{simul_data}}.}
}
\value{
A data frame with 4 columns (intercept estimate and standard error, and slope estimate and standard error), and as many rows as taxa.
}
\description{
Fit lump model
}
