% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_pglmm}
\alias{fit_pglmm}
\title{Fit Bayesian phylogenetic mixed model (brms)}
\usage{
fit_pglmm(simdata, delta = 0.95)
}
\arguments{
\item{simdata}{List generated by \code{\link{simul_data}}.}

\item{delta}{adapt_delta parameter for brm}
}
\value{
A data frame with 2 columns (intercept and slope estimate), and as many rows as taxa.
}
\description{
Fit Bayesian phylogenetic mixed model (brms)
}
