% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_niches.R
\name{plot_niches}
\alias{plot_niches}
\title{Plot niches}
\usage{
plot_niches(df = simdata$data2model, suitab.column = "suitab.invlogit")
}
\arguments{
\item{df}{Data frame}

\item{suitab.column}{Character. Name of the column containing suitability values}
}
\value{
A ggplot object
}
\description{
Plot niches
}
