% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_data.R
\name{simul_data}
\alias{simul_data}
\title{Simulate dataset}
\usage{
simul_data(nspp = NULL, nsite = NULL, seed = NULL, min.K = NULL)
}
\arguments{
\item{nspp}{Number of taxa}

\item{nsite}{Number of sites}

\item{seed}{Random seed}

\item{min.K}{Numeric Force minimum phylogenetic signal (Blomberg's K) for simulated intercepts and slopes?}
}
\value{
A dataframe
}
\description{
Simulate dataset
}
