% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_sims}
\alias{run_sims}
\title{Run a number of simulations with fixed parameters}
\usage{
run_sims(nsim, nspp, nsite, delta)
}
\arguments{
\item{nsim}{Number of replicate simulations to run}

\item{delta}{'adapt_delta' parameter for brm}
}
\value{
A data.frame with nrow = nspp*nsim
}
\description{
Run a number of simulations with fixed parameters
}
