% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run simulation}
\usage{
run_simulation(nspp, nsite, delta = 0.95)
}
\arguments{
\item{delta}{'adapt_delta' parameter for brm}
}
\value{
A data.frame
}
\description{
Run simulation
}
